refactoring, deletion of duplicates & unused files (tables.py)
