+++ .gitignore
@@ -9,3 +9,4 @@ logs
 tests/papers_rag/annett-o-test(2).owl
 .venv/
 data/annett-o-test.owl
+data/*
