+++ requirements/requirements-onnx.txt
@@ -1,2 +1,4 @@
+onnx
+fire
 sqlalchemy
 pgvector
