+++ tests/papers_rag/instantiate_annetto.py
@@ -4,7 +4,7 @@ import hashlib
 from datetime import datetime
 import time
 from typing import Dict, Any, Union, List, Optional
-from utils.onnx_additions.add_onnx import OnnxAddition
+from utils.fetch_onnx_db import OnnxAddition
 
 from owlready2 import Ontology, ThingClass, Thing, ObjectProperty, get_ontology
 from rapidfuzz import process, fuzz
